Move wallet-context to sui-sdk
