+++ crates/sui-cluster-test/src/cluster.rs
@@ -5,7 +5,6 @@ use async_trait::async_trait;
 use clap::*;
 use std::net::SocketAddr;
 use std::path::Path;
-use sui::client_commands::WalletContext;
 use sui_config::genesis_config::GenesisConfig;
 use sui_config::Config;
 use sui_config::SUI_KEYSTORE_FILENAME;
@@ -13,6 +12,7 @@ use sui_indexer::test_utils::start_test_indexer;
 use sui_indexer::IndexerConfig;
 use sui_keys::keystore::{AccountKeystore, FileBasedKeystore, Keystore};
 use sui_sdk::sui_client_config::{SuiClientConfig, SuiEnv};
+use sui_sdk::wallet_context::WalletContext;
 use sui_swarm::memory::Swarm;
 use sui_types::base_types::SuiAddress;
 use sui_types::crypto::KeypairTraits;
