+++ crates/sui-cluster-test/src/test_case/shared_object_test.rs
@@ -3,8 +3,8 @@
 
 use crate::{helper::ObjectChecker, TestCaseImpl, TestContext};
 use async_trait::async_trait;
-use sui::client_commands::WalletContext;
 use sui_json_rpc_types::{SuiExecutionStatus, SuiTransactionBlockEffectsAPI};
+use sui_sdk::wallet_context::WalletContext;
 use sui_types::object::Owner;
 use test_utils::transaction::{increment_counter, publish_basics_package_and_make_counter};
 use tracing::info;
