+++ crates/sui-faucet/src/faucet/simple_faucet.rs
@@ -13,12 +13,12 @@ use std::collections::HashSet;
 use std::path::Path;
 use typed_store::Map;
 
-use sui::client_commands::WalletContext;
 use sui_json_rpc_types::{
     SuiObjectDataOptions, SuiTransactionBlockEffectsAPI, SuiTransactionBlockResponse,
     SuiTransactionBlockResponseOptions,
 };
 use sui_keys::keystore::AccountKeystore;
+use sui_sdk::wallet_context::WalletContext;
 use sui_types::object::Owner;
 use sui_types::{
     base_types::{ObjectID, SuiAddress, TransactionDigest},
@@ -597,6 +597,7 @@ impl Faucet for SimpleFaucet {
 mod tests {
     use sui::client_commands::{SuiClientCommandResult, SuiClientCommands};
     use sui_json_rpc_types::SuiExecutionStatus;
+    use sui_sdk::wallet_context::WalletContext;
     use test_utils::network::TestClusterBuilder;
 
     use super::*;
