+++ crates/sui-sdk/Cargo.toml
@@ -9,6 +9,8 @@ edition = "2021"
 [dependencies]
 anyhow = "1.0.64"
 async-trait = "0.1.61"
+clap = { version = "3.2.17", features = ["derive"] }
+colored = "2.0.0"
 jsonrpsee = { version = "0.16.2", features = ["http-client"] }
 serde = { version = "1.0.144", features = ["derive"] }
 serde_with = { version = "2.1.0", features = ["hex"] }
@@ -28,6 +30,8 @@ sui-json = { path = "../sui-json" }
 sui-keys = { path = "../sui-keys" }
 sui-config = { path = "../sui-config" }
 shared-crypto = { path = "../shared-crypto" }
+tracing = "0.1.36"
+
 move-core-types.workspace = true
 
 fastcrypto.workspace = true
