+++ crates/sui/src/sui_commands.rs
@@ -27,7 +27,7 @@ use sui_keys::keystore::{AccountKeystore, FileBasedKeystore, Keystore};
 use sui_swarm::memory::Swarm;
 use sui_types::crypto::{SignatureScheme, SuiKeyPair};
 
-use crate::client_commands::{SuiClientCommands, WalletContext};
+use crate::client_commands::SuiClientCommands;
 use crate::console::start_console;
 use crate::fire_drill::{run_fire_drill, FireDrill};
 use crate::genesis_ceremony::{run, Ceremony};
@@ -35,6 +35,7 @@ use crate::keytool::KeyToolCommand;
 use crate::validator_commands::SuiValidatorCommand;
 use sui_move::{self, execute_move_command};
 use sui_sdk::sui_client_config::{SuiClientConfig, SuiEnv};
+use sui_sdk::wallet_context::WalletContext;
 
 #[allow(clippy::large_enum_variant)]
 #[derive(Parser)]
