+++ crates/sui/tests/snapshot_tests.rs
@@ -3,8 +3,9 @@
 
 use clap::Parser;
 use insta::assert_json_snapshot;
-use sui::{client_commands::WalletContext, sui_commands::SuiCommand};
+use sui::sui_commands::SuiCommand;
 use sui_macros::sim_test;
+use sui_sdk::wallet_context::WalletContext;
 use test_utils::network::TestClusterBuilder;
 
 // special constants for substitution in commands
